+++ internal/provider/provider.go
@@ -281,6 +281,7 @@ func (p *OMVProvider) Resources(_ context.Context) []func() resource.Resource {
 		NewWorkbenchSettingsResource,
 		NewSSLCertificateResource,
 		NewSSHCertificateResource,
+		NewSMBShareResource,
 	}
 }